+++ src/test/java/team/serenity/logic/commands/studentinfo/FlagAttCommandTest.java
@@ -1,0 +1,7 @@
+package team.serenity.logic.commands.studentinfo;
+
+import static org.junit.jupiter.api.Assertions.*;
+
+class FlagAttCommandTest {
+
+}
