+++ src/test/java/team/serenity/logic/commands/studentinfo/MarkAbsentCommandTest.java
@@ -152,118 +152,3 @@ class MarkAbsentCommandTest {
     }
 }
 
-/**
- * A Model stub containing absent students
- */
-class ModelStubWithStudentsPresent extends ModelStub {
-    private Group uniqueGroup;
-    private Lesson uniqueLesson;
-
-
-    @Override
-    public ObservableList<Group> getFilteredGroupList() {
-        List<Group> grpList = new ArrayList<>();
-        this.uniqueGroup = new GroupBuilder().build();
-        grpList.add(uniqueGroup);
-        UniqueList<Group> groupUniqueList = new UniqueGroupList();
-        groupUniqueList.setElementsWithList(grpList);
-        return groupUniqueList.asUnmodifiableObservableList();
-    }
-
-    @Override
-    public ObservableList<Lesson> getFilteredLessonList() {
-        List<Lesson> lsnList = new ArrayList<>();
-        this.uniqueLesson = new LessonBuilder().build();
-        lsnList.add(uniqueLesson);
-        UniqueList<Lesson> lessonUniqueList = new UniqueLessonList();
-        lessonUniqueList.setElementsWithList(lsnList);
-        return lessonUniqueList.asUnmodifiableObservableList();
-    }
-
-    @Override
-    public UniqueList<StudentInfo> getListOfStudentsInfoFromGroupAndLesson(Group group, Lesson lesson) {
-        GroupLessonKey key = new GroupLessonKey(group.getGroupName(), lesson.getLessonName());
-        GroupLessonKey mapKey = new GroupLessonKey(uniqueGroup.getGroupName(), uniqueLesson.getLessonName());
-        Map<GroupLessonKey, UniqueList<StudentInfo>> uniqueStudentInfoList = new HashMap<>();
-        UniqueList<StudentInfo> studentInfoList = uniqueLesson.getStudentsInfo();
-        ObservableList<StudentInfo> observableStudentInfo = studentInfoList.asUnmodifiableObservableList();
-        for (int i = 0; i < studentInfoList.size(); i++) {
-            StudentInfo studentInfo = observableStudentInfo.get(i);
-            Attendance updatedAttendance = studentInfo.getAttendance().setNewAttendance(true);
-            studentInfo.updateAttendance(updatedAttendance);
-        }
-        uniqueStudentInfoList.put(mapKey, uniqueLesson.getStudentsInfo());
-        return uniqueStudentInfoList.get(key);
-    }
-
-    @Override
-    public void updateStudentsInfoList() {
-        return;
-    }
-
-    @Override
-    public void updateLessonList() {
-        return;
-    }
-
-}
-
-/**
- * A Model stub containing an absent student
- */
-class ModelStubWithIndexPresent extends ModelStub {
-    private Group uniqueGroup;
-    private Lesson uniqueLesson;
-
-    @Override
-    public ObservableList<Group> getFilteredGroupList() {
-        List<Group> grpList = new ArrayList<>();
-        this.uniqueGroup = new GroupBuilder().withName("G01")
-                .withStudents(new Student("Aaron Tan", "A0123456U"))
-                .withClasses("1-1").build();
-        grpList.add(uniqueGroup);
-        UniqueList<Group> groupUniqueList = new UniqueGroupList();
-        groupUniqueList.setElementsWithList(grpList);
-        return groupUniqueList.asUnmodifiableObservableList();
-    }
-
-    @Override
-    public ObservableList<Lesson> getFilteredLessonList() {
-        List<Lesson> lsnList = new ArrayList<>();
-        this.uniqueLesson = new LessonBuilder()
-                .withName("1-1")
-                .withStudentInfos(new StudentInfo(new Student("Aaron Tan", "A0123456U")))
-                .build();
-        lsnList.add(uniqueLesson);
-        UniqueList<Lesson> lessonUniqueList = new UniqueLessonList();
-        lessonUniqueList.setElementsWithList(lsnList);
-        return lessonUniqueList.asUnmodifiableObservableList();
-    }
-
-    @Override
-    public UniqueList<StudentInfo> getListOfStudentsInfoFromGroupAndLesson(Group group, Lesson lesson) {
-        GroupLessonKey key = new GroupLessonKey(group.getGroupName(), lesson.getLessonName());
-        GroupLessonKey mapKey = new GroupLessonKey(uniqueGroup.getGroupName(), uniqueLesson.getLessonName());
-        Map<GroupLessonKey, UniqueList<StudentInfo>> uniqueStudentInfoList = new HashMap<>();
-        UniqueList<StudentInfo> studentInfoList = uniqueLesson.getStudentsInfo();
-        ObservableList<StudentInfo> observableStudentInfo = studentInfoList.asUnmodifiableObservableList();
-        for (int i = 0; i < studentInfoList.size(); i++) {
-            StudentInfo studentInfo = observableStudentInfo.get(i);
-            Attendance updatedAttendance = studentInfo.getAttendance().setNewAttendance(true);
-            studentInfo.updateAttendance(updatedAttendance);
-        }
-        uniqueStudentInfoList.put(mapKey, uniqueLesson.getStudentsInfo());
-        return uniqueStudentInfoList.get(key);
-    }
-
-    @Override
-    public void updateLessonList() {
-        return;
-    }
-
-    @Override
-    public void updateStudentsInfoList() {
-        return;
-    }
-
-}
