+++ src/test/java/team/serenity/logic/commands/studentinfo/ModelStubWithStudentsPresent.java
@@ -1,0 +1,75 @@
+package team.serenity.logic.commands.studentinfo;
+
+import java.util.ArrayList;
+import java.util.HashMap;
+import java.util.List;
+import java.util.Map;
+
+import javafx.collections.ObservableList;
+import team.serenity.model.group.Group;
+import team.serenity.model.group.GroupLessonKey;
+import team.serenity.model.group.UniqueGroupList;
+import team.serenity.model.group.lesson.Lesson;
+import team.serenity.model.group.lesson.UniqueLessonList;
+import team.serenity.model.group.studentinfo.Attendance;
+import team.serenity.model.group.studentinfo.StudentInfo;
+import team.serenity.model.util.UniqueList;
+import team.serenity.testutil.GroupBuilder;
+import team.serenity.testutil.LessonBuilder;
+import team.serenity.testutil.ModelStub;
+
+/**
+ * A Model stub containing absent students
+ */
+class ModelStubWithStudentsPresent extends ModelStub {
+    private Group uniqueGroup;
+    private Lesson uniqueLesson;
+
+
+    @Override
+    public ObservableList<Group> getFilteredGroupList() {
+        List<Group> grpList = new ArrayList<>();
+        this.uniqueGroup = new GroupBuilder().build();
+        grpList.add(uniqueGroup);
+        UniqueList<Group> groupUniqueList = new UniqueGroupList();
+        groupUniqueList.setElementsWithList(grpList);
+        return groupUniqueList.asUnmodifiableObservableList();
+    }
+
+    @Override
+    public ObservableList<Lesson> getFilteredLessonList() {
+        List<Lesson> lsnList = new ArrayList<>();
+        this.uniqueLesson = new LessonBuilder().build();
+        lsnList.add(uniqueLesson);
+        UniqueList<Lesson> lessonUniqueList = new UniqueLessonList();
+        lessonUniqueList.setElementsWithList(lsnList);
+        return lessonUniqueList.asUnmodifiableObservableList();
+    }
+
+    @Override
+    public UniqueList<StudentInfo> getListOfStudentsInfoFromGroupAndLesson(Group group, Lesson lesson) {
+        GroupLessonKey key = new GroupLessonKey(group.getGroupName(), lesson.getLessonName());
+        GroupLessonKey mapKey = new GroupLessonKey(uniqueGroup.getGroupName(), uniqueLesson.getLessonName());
+        Map<GroupLessonKey, UniqueList<StudentInfo>> uniqueStudentInfoList = new HashMap<>();
+        UniqueList<StudentInfo> studentInfoList = uniqueLesson.getStudentsInfo();
+        ObservableList<StudentInfo> observableStudentInfo = studentInfoList.asUnmodifiableObservableList();
+        for (int i = 0; i < studentInfoList.size(); i++) {
+            StudentInfo studentInfo = observableStudentInfo.get(i);
+            Attendance updatedAttendance = studentInfo.getAttendance().setNewAttendance(true);
+            studentInfo.updateAttendance(updatedAttendance);
+        }
+        uniqueStudentInfoList.put(mapKey, uniqueLesson.getStudentsInfo());
+        return uniqueStudentInfoList.get(key);
+    }
+
+    @Override
+    public void updateStudentsInfoList() {
+        return;
+    }
+
+    @Override
+    public void updateLessonList() {
+        return;
+    }
+
+}
